pyproject.toml
[tool.poetry]
name = "magfest"
version = "0.1.0"
description = ""
authors = ["Nate Shoffner <nate.shoffner@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.9.2"
requests = "^2.32.3"
twilio = "^9.3.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
